# src/lib/admin/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tree-gateway",
        "description": "The Tree Gateway Administration API",
        "version": "1.0.0"
    },
    "host": "tree-gateway.com",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "basePath": "/",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/apis/{apiId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve the API configuration for the given id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Api"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the API with the given id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "apiData",
                        "description": "API data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Api"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Invalid API",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API Identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "API cannot be deleted",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
...
        "/apis/{apiId}/throttling": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of throttling configuration.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Throttling"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new throttling configuration.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Throttling data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Throttling"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API not found"
                    }
                }
            }
        },
        "/apis/{apiId}/throttling/{throttlingId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a throttling configuration by ID.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "throttlingId",
                        "description": "Throttling ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Throttling"
                        }
                    },
                    "404": {
                        "description": "API/Throttling not found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a throttling configuration",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "throttlingId",
                        "description": "Throttling ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Throttling config",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Throttling"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Invalid throttling config",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API/Throttling not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a throttling configuration",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apiId",
                        "description": "API identifier",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "throttlingId",
                        "description": "Throttling ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cannot delete throttling config",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "API/Throttling not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Api": {
            "type": "object",
            "description": "API configuration",
            "properties": {
                "id": {
                    "type": "number",
                    "description": "The id unique identifier"
                },
                "name": {
                    "type": "string",
                    "description": "API name"
                },
                "description": {
                    "type": "string",
                    "description": "API description"
                },
                "url": {
                    "type": "string",
                    "description": "The URL for this API"
                }
            }
        },
...
        "Throttling": {
            "type": "object",
            "description": "Throttling configuration",
            "properties": {
                "windowMs": {
                    "type": "integer",
                    "description": "How long to keep records of requests in memor (milliseconds)."
                },
                "delayAfter": {
                    "type": "integer",
                    "description": "Max number of connections during windowMs before starting to delay responses. Defaults to 1. Set to 0 to disable delaying."
                },
                "delayMs": {
                    "type": "integer",
                    "description": "How long to delay the response, multiplied by (number of recent hits - delayAfter). Defaults to 1000 (1 second). Set to 0 to disable delaying."
                },
                "max": {
                    "type": "integer",
                    "description": "Max number of connections during windowMs milliseconds before sending a 429 response. Defaults to 5. Set to 0 to disable."
                },
                "message": {
                    "type": "string",
                    "description": "Error message returned when max is exceeded. Defaults to 'Too many requests, please try again later.'"
                },
                "statusCode": {
                    "type": "integer",
                    "description": "HTTP status code returned when max is exceeded. Defaults to 429."
                },
                "headers": {
                    "type": "boolean",
                    "description": "Enable header to show request limit and current usage."
                },
                "keyGenerator": {
                    "type": "string",
                    "description": "The name of the function used to generate keys. By default user IP address (req.ip) is used."
                },
                "handler": {
                    "type": "string",
                    "description": "The name of the function to execute once the max limit is exceeded. It receives the request and the response objects. The \"next\" param is available if you need to pass to the next middleware."
                },
                "groups": {
                    "type": "array",
                    "description": "A list of groups that should be handled by this limiter. If not provided, everything will be handled. Defaults to *.",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Error code."
                },
                "message": {
                    "type": "string",
                    "description": "Error message"
                }
            }
        }
    }
}